Refactor activity log parser
