+++ app/api/activitylog/activitylogParser.js
@@ -1,75 +1,14 @@
 import templates from 'api/templates';
 import entities from 'api/entities';
-import semanticSearchModel from 'api/semanticsearch/model';
-import { allLanguages } from 'shared/languagesList';
 
-const methods = {
-  create: 'CREATE',
-  update: 'UPDATE',
-  delete: 'DELETE'
-};
-
-const formatLanguage = (langKey) => {
-  const lang = allLanguages.find(({ key }) => key === langKey);
-  return lang ? `${lang.label} (${lang.key})` : langKey;
-};
-
-const generateCreateUpdateBeautifier = (resourceName, idField, nameField) => async (log) => {
-  const data = JSON.parse(log.body);
-  const name = data[nameField];
-
-  const semantic = {
-    beautified: true,
-    name
-  };
-
-  if (data[idField]) {
-    semantic.name = name ? `${name} (${data[idField]})` : `${data[idField]}`;
-    semantic.action = methods.update;
-    semantic.description = `Updated ${resourceName}`;
-  } else {
-    semantic.action = methods.create;
-    semantic.description = `Created ${resourceName}`;
-  }
-
-  return semantic;
-};
-
-
-const generateDeleteBeautifier = (resourceName, idField) => async (log) => {
-  const data = JSON.parse(log.query);
-
-  return {
-    beautified: true,
-    action: methods.delete,
-    description: `Deleted ${resourceName}`,
-    name: data[idField]
-  };
-};
-
-const generatePlainDescriptionBeautifier = (description, action = methods.update) => async () => ({
-  beautified: true,
-  action,
-  description
-});
-
-const generateSemanticSearchUpdateBeautifier = description => async (log) => {
-  const data = JSON.parse(log.body);
-  const search = await semanticSearchModel.getById(data.searchId);
-
-  const semantic = {
-    beautified: true,
-    action: methods.update,
-    description,
-    name: data.searchId
-  };
-
-  if (search) {
-    semantic.name = `${search.searchTerm} (${data.searchId})`;
-  }
-
-  return semantic;
-};
+import {
+  methods,
+  generateCreateUpdateBeautifier,
+  generateDeleteBeautifier,
+  generatePlainDescriptionBeautifier,
+  generateSemanticSearchUpdateBeautifier,
+  formatLanguage
+} from './helpers';
 
 const entitiesPOST = async (log) => {
   const data = JSON.parse(log.body);
